+++ app/editor/components/CommandMenu.tsx
@@ -11,8 +11,7 @@ import { CommandFactory } from "@shared/editor/lib/Extension";
 import filterExcessSeparators from "@shared/editor/lib/filterExcessSeparators";
 import { EmbedDescriptor, MenuItem } from "@shared/editor/types";
 import { depths } from "@shared/styles";
-import { supportedImageMimeTypes } from "@shared/utils/files";
-import { getEventFiles } from "@shared/utils/getDataTransferFiles";
+import { supportedImageMimeTypes, getEventFiles } from "@shared/utils/files";
 import Scrollable from "~/components/Scrollable";
 import { Dictionary } from "~/hooks/useDictionary";
 import Input from "./Input";
